% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{adm_table_toy}
\alias{adm_table_toy}
\title{Toy example of admixture table.}
\format{An object of class \code{data.frame} with 1 rows and 4 columns.}
\usage{
adm_table_toy
}
\description{
Toy example of admixture table.
}
\keyword{datasets}
