% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{allele_specific_cna_table_toy}
\alias{allele_specific_cna_table_toy}
\title{Toy example of allele specific table of somatic copy number.}
\format{An object of class \code{data.frame} with 4 rows and 15 columns.}
\usage{
allele_specific_cna_table_toy
}
\description{
Toy example of allele specific table of somatic copy number.
}
\keyword{datasets}
