% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{bt_toy}
\alias{bt_toy}
\title{Toy example of beta table.}
\format{An object of class \code{data.frame} with 4 rows and 10 columns.}
\usage{
bt_toy
}
\description{
Toy example of beta table.
}
\keyword{datasets}
