% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{error_table}
\alias{error_table}
\title{Beta estimation error.}
\format{A data frame column names mean.cov, n.info.snps, and
adm.estimation.error \describe{ \item{mean.cov}{genomic segment coverage}
\item{n.info.snps}{number of informative SNPs}
\item{adm.estimation.error}{expected error on beta estimate } }}
\usage{
error_table
}
\description{
A precomputed table reporting for different combinations of coverage and
number of informative SNPs  the expected error of the  beta value computed by
function \code{\link[CLONETv2:compute_beta_table]{compute_beta_table}}.
}
\keyword{datasets}
