% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{pileup_normal_toy}
\alias{pileup_normal_toy}
\title{Toy example of normal pileup data.}
\format{An object of class \code{data.frame} with 816 rows and 11 columns.}
\usage{
pileup_normal_toy
}
\description{
Toy example of normal pileup data.
}
\keyword{datasets}
