% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{pileup_tumor_toy}
\alias{pileup_tumor_toy}
\title{Toy example of tumor pileup data.}
\format{An object of class \code{data.frame} with 816 rows and 11 columns.}
\usage{
pileup_tumor_toy
}
\description{
Toy example of tumor pileup data.
}
\keyword{datasets}
