% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{pl_table_toy}
\alias{pl_table_toy}
\title{Toy example of ploidy table.}
\format{An object of class \code{data.frame} with 1 rows and 2 columns.}
\usage{
pl_table_toy
}
\description{
Toy example of ploidy table.
}
\keyword{datasets}
