% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{scna_clonality_table_toy}
\alias{scna_clonality_table_toy}
\title{Toy example of clonality table of somatic copy number.}
\format{An object of class \code{data.frame} with 4 rows and 25 columns.}
\usage{
scna_clonality_table_toy
}
\description{
Toy example of clonality table of somatic copy number.
}
\keyword{datasets}
