% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{seg_tb_toy}
\alias{seg_tb_toy}
\title{Toy example of segmetd data.}
\format{An object of class \code{data.frame} with 4 rows and 6 columns.}
\usage{
seg_tb_toy
}
\description{
Toy example of segmetd data.
}
\keyword{datasets}
