% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{snv_clonality_table_toy}
\alias{snv_clonality_table_toy}
\title{Toy example of snv clonality table.}
\format{An object of class \code{data.frame} with 2 rows and 78 columns.}
\usage{
snv_clonality_table_toy
}
\description{
Toy example of snv clonality table.
}
\keyword{datasets}
