% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLONETv2_data.R
\docType{data}
\name{snv_reads_toy}
\alias{snv_reads_toy}
\title{Toy example of snv data.}
\format{An object of class \code{data.frame} with 2 rows and 71 columns.}
\usage{
snv_reads_toy
}
\description{
Toy example of snv data.
}
\keyword{datasets}
